+++ dal-cpp/dal/math/random/sobol.hpp
@@ -20,7 +20,7 @@ polish is boolean
 -IF-------------------------------------------------------------------------*/
 
 namespace Dal {
-    SequenceSet_* NewSobol(int size, size_t iPath, bool precise = false, bool polish = false);
+    SequenceSet_* NewSobol(int size, size_t iPath, bool precise = true, bool polish = false);
 
     class BASE_EXPORT SobolRSG_: public Storable_ {
         std::unique_ptr<SequenceSet_> rsg_;
@@ -29,7 +29,7 @@ namespace Dal {
         bool precise_;
         bool polish_;
     public:
-        SobolRSG_(const String_& name, double iPath, double nDim = 1, bool precise = false, bool polish = false)
+        SobolRSG_(const String_& name, double iPath, double nDim = 1, bool precise = true, bool polish = false)
             : Storable_("SobolRSG", name), i_path_(iPath), ndim_(nDim), precise_(precise), polish_(polish) {
             rsg_.reset(NewSobol(static_cast<int>(ndim_), static_cast<size_t>(i_path_), precise, polish));
         }